updated consumer retry, failure message send to topic and db

+++ device-event-producer/src/main/java/com/kafka/device/domain/DeviceEvent.java
@@ -1,7 +1,7 @@
 package com.kafka.device.domain;
 
 public record DeviceEvent(
-        Long deviceEventId,
+        Integer deviceEventId,
         DeviceEventType deviceEventType,
         Device device
